sw/ [FEATURE]: add programs for interaction with a card
The pkt_{receive,transmit}.c programs are meant to interact with the DMA
engine to facilitate data transmission. The gls_mod.py allows to measure
a throughput of a DMA with various packet length. The program_fpga.tcl
simplifies the programming of the FPGA where the user does not have to
open thi Vivado GUI and everything can be done in a terminal.